Split front page
Split front page into multiple files for easier management.
