+++ .babelrc
@@ -1,4 +1,13 @@
 {
-  "presets": ["es2015", "stage-0", "react"],
-  "plugins": ["syntax-async-functions", "transform-decorators-legacy"]
+	"presets": [
+		"es2015",
+		"stage-0",
+		"react"
+	],
+	"plugins": [
+		"syntax-async-functions",
+		"transform-regenerator",
+		"transform-decorators-legacy",
+		"react-hot-loader/babel"
+	]
 }
