+++ client/react/main.js
@@ -13,6 +13,7 @@ import { verifySession } from '../redux/actions/session.js';
 // MaterialUI Theme
 import baseTheme from 'material-ui/styles/baseThemes/lightBaseTheme';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
+
 const muiTheme = getMuiTheme(baseTheme);
 
 import AppRouter from './AppRouter.js';
@@ -34,9 +35,9 @@ class App extends React.Component {
 	
 	// This is needed for MUI.
 	getChildContext = () => {
-		return {muiTheme};
+		return { muiTheme };
 	};
-
+	
 	componentWillMount() {
 		if (this.props.inSession === -1) {
 			this.props.verifySession(true);
@@ -45,16 +46,16 @@ class App extends React.Component {
 	
 	render() {
 		return this.props.inSession === -1 ? (
-			<div>
+			<div style={{ width: '100%', height: '100%' }}>
 				<Loading />
 				<Success />
 			</div>
 		) : (
-			<div>
+			<div style={{ width: '100%', height: '100%' }}>
 				<Loading />
 				<Success />
 				
-				<AppRouter/>
+				<AppRouter />
 			</div>
 		);
 	}
@@ -63,15 +64,17 @@ class App extends React.Component {
 /** Routes */
 const MasterRouter = () => (
 	<ConnectedRouter history={browserHistory}>
-		<div style={{top: 0, left: 0, width: '100vw', height: '97.5vh', transform: 'translateX(-1vw) translateY(-1.5vh)'}}>
-			<Route path="/" component={App}/>
+		<div style={{ width: '100vw', height: '100vh' }}>
+			<Route path="/" component={App} />
 		</div>
 	</ConnectedRouter>
 );
 
 const Main = (
 	<Provider store={store}>
-		<MasterRouter/>
+		<div style={{ overflow: 'hidden' }}>
+			<MasterRouter />
+		</div>
 	</Provider>
 );
