+++ package.json
@@ -4,23 +4,11 @@
 	"description": "A web application that connects mentors with mentees for peer tutoring.",
 	"main": "server/entry.js",
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
+		"test": "jest",
 		"dev:server": "nodemon --exec ts-node -- ./server/entry.ts",
 		"dev:client": "node client/devServerEntry",
 		"dev": "npm-run-all --parallel dev:*",
 		"build": "gulp server && webpack"
-	},
-	"babel": {
-		"presets": [
-			"es2015",
-			"stage-0",
-			"react"
-		],
-		"plugins": [
-			"syntax-async-functions",
-			"transform-regenerator",
-			"react-hot-loader/babel"
-		]
 	},
 	"repository": {
 		"type": "git",
@@ -49,7 +37,7 @@
 		"js-cookie": "^2.1.4",
 		"jsonwebtoken": "^7.4.1",
 		"lodash": "^4.17.4",
-		"material-ui": "^0.18.3",
+		"material-ui": "^0.18.7",
 		"mysql": "^2.13.0",
 		"mysql2": "^1.3.5",
 		"nodemailer": "^4.0.1",
@@ -59,11 +47,13 @@
 		"react": "^15.6.1",
 		"react-dom": "^15.6.1",
 		"react-hot-loader": "^3.0.0-beta.7",
+		"react-jss": "^7.0.2",
 		"react-redux": "^5.0.5",
 		"react-router": "^4.0.0",
 		"react-router-redux": "5.0.0-alpha.6",
 		"react-tap-event-plugin": "^2.0.1",
 		"redux": "^3.6.0",
+		"redux-form": "^7.0.3",
 		"redux-logger": "^3.0.6",
 		"redux-thunk": "^2.2.0",
 		"sequelize": "^4.1.0",
@@ -84,6 +74,7 @@
 		"@types/nodemailer": "^1.3.33",
 		"@types/randomstring": "^1.1.6",
 		"@types/sequelize": "^4.0.65",
+		"babel-jest": "^20.0.3",
 		"babel-loader": "^7.0.0",
 		"babel-plugin-syntax-async-functions": "^6.13.0",
 		"babel-plugin-transform-regenerator": "^6.24.1",
@@ -93,6 +84,7 @@
 		"compression-webpack-plugin": "^0.4.0",
 		"gulp": "^3.9.1",
 		"gulp-typescript": "^3.2.1",
+		"jest": "^20.0.4",
 		"npm-run-all": "^4.0.2",
 		"path": "^0.12.7",
 		"ts-loader": "^2.3.1",
